+++ Assets/Scripts/Objects Managment/MPD Tools/Valve.cs
@@ -8,7 +8,7 @@ public class Valve : BaseTool, IClosable
     [RangeAttribute(0, 100)]
     public float Openness;
 
-    public double InnerDiameter=9;
+    public float InnerDiameter=9;
     public bool IsOpen;
     
 
@@ -32,16 +32,16 @@ public class Valve : BaseTool, IClosable
             this.Openness += 5f;
     }
 
-    public override string HeadLossEq(string param, double flowRate)
+    public override string HeadLossEq(string param, float flowRate)
     {
    
         if (Openness <= 0)
         {
             return ("0");
         }
-        //double velocity = 4 * flowRate / (Math.PI * InnerDiameter * InnerDiameter);
-        //double rN = Fluids.Reynolds(Fluids.Density, velocity, InnerDiameter, Fluids.Viscosity);
-        //double f = Fluids.FrictionFactor(rN, Roughness);
+        //float velocity = 4 * flowRate / (Math.PI * InnerDiameter * InnerDiameter);
+        //float rN = Fluids.Reynolds(Fluids.Density, velocity, InnerDiameter, Fluids.Viscosity);
+        //float f = Fluids.FrictionFactor(rN, Roughness);
         //return (8 * f * Length / (Math.PI * Math.PI * 9.81 * Math.Pow(InnerDiameter, 5))).ToString("f5") + "*" + param + "^2";
         return "0";
     }
